+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.6.8] - 2018-06-25
 
 ### Changed
 
@@ -136,7 +136,8 @@ This is a major release with some API changes.
 
 
 
-[Unreleased]: https://github.com/sdispater/cleo/compare/0.6.7...master
+[Unreleased]: https://github.com/sdispater/cleo/compare/0.6.8...master
+[0.6.8]: https://github.com/sdispater/cleo/releases/tag/0.6.8
 [0.6.7]: https://github.com/sdispater/cleo/releases/tag/0.6.7
 [0.6.6]: https://github.com/sdispater/cleo/releases/tag/0.6.6
 [0.6.5]: https://github.com/sdispater/cleo/releases/tag/0.6.5